+++ src/main/java/net/minecraft/item/ItemSword.java
@@ -13,7 +13,7 @@ import net.minecraft.world.World;
 
 public class ItemSword extends Item
 {
-    private float field_150934_a;
+    public float field_150934_a;
     private final Item.ToolMaterial field_150933_b;
     private static final String __OBFID = "CL_00000072";